iio: ad9680: fix incompatible pointer for get/set callbacks
Error is:
```
drivers/iio/adc/ad9680.c:503:9: error: initialization from incompatible pointer type [-Werror=incompatible-pointer-types]
  .set = ad9680_testmode_write,
         ^
drivers/iio/adc/ad9680.c:503:9: note: (near initialization for 'ad9680_testmode_enum.set')
drivers/iio/adc/ad9680.c:504:9: error: initialization from incompatible pointer type [-Werror=incompatible-pointer-types]
  .get = ad9680_testmode_read,

```

Though, this is a trivial error [that could be downgraded
to a warning], some more [pedantic] compiler options could
error here [as it happened in the case above].

+++ drivers/iio/adc/ad9680.c
@@ -464,7 +464,7 @@ static int ad9680_set_scale(struct axiadc_converter *conv, int val, int val2)
 	return -EINVAL;
 }
 
-static ssize_t ad9680_testmode_read(struct iio_dev *indio_dev,
+static int ad9680_testmode_read(struct iio_dev *indio_dev,
 	const struct iio_chan_spec *chan)
 {
 	struct axiadc_converter *conv = iio_device_get_drvdata(indio_dev);
@@ -472,7 +472,7 @@ static ssize_t ad9680_testmode_read(struct iio_dev *indio_dev,
 	return conv->testmode[chan->channel];
 }
 
-static ssize_t ad9680_testmode_write(struct iio_dev *indio_dev,
+static int ad9680_testmode_write(struct iio_dev *indio_dev,
 	const struct iio_chan_spec *chan, unsigned int item)
 {
 	int ret;
